+++ src/js/api.ts
@@ -692,6 +692,14 @@ export class PyodideAPI {
   static get lockfile() {
     return API.lockfile;
   }
+
+  /**
+   * Returns the base URL of the lockfile, which is used to locate the packages
+   * distributed with the lockfile.
+   */
+  static get lockfileBaseUrl() {
+    return API.lockfileBaseUrl;
+  }
 }
 
 /** @hidden */