+++ docs/project/changelog.md
@@ -35,6 +35,11 @@ myst:
 - {{ Fix }} Add the current working directory to the path instead of `$HOME`.
   {pr}`5630`
 
+- {{ Breadking }} When `lockfileURL` is given to `loadPyodide`, the
+  base URL for the packages is now calculated from the lockfile URL, not from
+  the `indexURL`.
+  {pr}`5652`
+
 ### `python` CLI entrypoint
 
 - {{ Fix }} The `python` CLI now mounts the `/tmp` directory. In